+++ testapps/dynTest.c
@@ -1,15 +1,16 @@
 #include <stdio.h>
+#include <stdint.h>
 
 #define DEFAULT_TRIP_COUNT 1234567
 
 int main(int argc, char* argv[]){
-    int i, n, k = 0;
+    int64_t i, n, k = 0;
 
 
     if (argc < 2){
         n = DEFAULT_TRIP_COUNT;
     } else {
-        sscanf(argv[1], "%d", &n);
+        sscanf(argv[1], "%lld", &n);
     }
 
     for (i = 0; i < n; i++){
@@ -20,7 +21,7 @@ int main(int argc, char* argv[]){
         }
     }
 
-    printf("Ran loop %d times -- %d\n", n, k);
+    printf("Ran loop %lld times -- %lld\n", n, k);
     printf("(Test Application Successfull)\n");
     fflush(stdout);
     return 0;
